Added FullText section. Added document index and filter. Edited page navigation to include Documents. Click on lexical unit in a paragraph to see the annotation (only the sentence). Static color list. Search page : removed tree node for documents (too slow to render). Changed initialisation : fetch type EAGER (for corpus.getDocuments()).

+++ src/main/java/jp/keio/jfn/wat/domain/Sentence.java
@@ -28,7 +28,7 @@ public class Sentence implements Serializable {
 
 	private Timestamp modifiedDate;
 
-	private byte paragraphOrder;
+	private int paragraphOrder;
 
 	private String text;
 
@@ -96,11 +96,11 @@ public class Sentence implements Serializable {
 		this.modifiedDate = modifiedDate;
 	}
 
-	public byte getParagraphOrder() {
+	public int getParagraphOrder() {
 		return this.paragraphOrder;
 	}
 
-	public void setParagraphOrder(byte paragraphOrder) {
+	public void setParagraphOrder(int paragraphOrder) {
 		this.paragraphOrder = paragraphOrder;
 	}
